requirements.txt
alembic==1.4.2
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.3
autopep8==1.5.4
bcrypt==3.2.0
blinker==1.4
convertdate==2.3.0
decorator==4.4.2
distlib==0.3.0
Flask==1.1.2
Flask-Dance==3.1.0
Flask-JWT-Extended==3.25.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.1
idna==2.10
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.8
sqlparse==0.3.0
urllib3==1.26.2
URLObject==2.4.3
virtualenv==20.0.20
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
mysqlclient==2.1.1
